Lots of changes, merged color and gabor
I tried a lot of different things to determine how to find the house but
I couldn't quite finish it. I have a location estimate for the house
(centroid) but we just need to link the Gabor edges with the Color
building. I can think of many ways just couldn't get one to work :(