Implemented the Voting matrix!
I was able to get the voting matrix work in descVec. I used dilation on
each color segmentation then did a logical and on the dilated color and
edge. Then I added all points up that over lapped to determine a vote.
The higher the vote the more points that are closer in edge to color